add userId param to methods

+++ storage/src/main/scala/chat/Doobie.scala
@@ -7,18 +7,19 @@ import doobie.util.transactor.Transactor
 import doobie.util.update.Update0
 import todo.Repository.Name
 import todo.TodoError.UnexpectedError
-import todo.{ChatID, NumberOfTask, Repository, TodoTask}
-import zio.{Has, IO, Task, ZIO, ZLayer}
+import todo.{ChatID, NumberOfTask, Repository, TodoTask, UserID}
+import zio.{Task}
 import zio.interop.catz._
 
 private[chat] final case class Doobie(xa: Transactor[Task]) extends Service {
   override def add(
       chatID: ChatID,
       name: Repository.Name,
-      numberOfTask: NumberOfTask
+      numberOfTask: NumberOfTask,
+      userID: UserID
   ): Task[Unit] =
     SQL
-      .create(chatID, name, numberOfTask)
+      .create(chatID, name, numberOfTask, userID)
       .withUniqueGeneratedKeys[Long]("Id")
       .transact(xa)
       .unit
@@ -26,10 +27,11 @@ private[chat] final case class Doobie(xa: Transactor[Task]) extends Service {
 
   override def remove(
       chatID: ChatID,
-      numberOfTask: NumberOfTask
+      numberOfTask: NumberOfTask,
+      userID: UserID
   ): Task[Unit] = {
     SQL
-      .delete(chatID, numberOfTask)
+      .delete(chatID, numberOfTask, userID)
       .run
       .transact(xa)
       .unit
@@ -60,33 +62,58 @@ private[chat] final case class Doobie(xa: Transactor[Task]) extends Service {
   override def update(
       chatID: ChatID,
       name: Name,
-      numberOfTask: NumberOfTask
+      numberOfTask: NumberOfTask,
+      userID: UserID
   ): Task[Unit] =
-    SQL.update(chatID, numberOfTask, name).run.transact(xa).unit.orDie
+    SQL.update(chatID, numberOfTask, name, userID).run.transact(xa).unit.orDie
 
+  override def listUserTasks(
+      chatID: ChatID,
+      userID: UserID
+  ): Task[Set[TodoTask]] =
+    SQL.getTasksByUser(chatID, userID).to[Set].transact(xa).orDie
 }
 
 private object SQL {
-  def create(chatID: ChatID, name: Name, numberOfTask: NumberOfTask): Update0 =
-    sql"""Insert INTO Task (Chat_ID, Task_Name, Ordering)
-        values (${chatID.value},${name.value}, ${numberOfTask.value})
+  def create(
+      chatID: ChatID,
+      name: Name,
+      numberOfTask: NumberOfTask,
+      userID: UserID
+  ): Update0 =
+    sql"""Insert INTO Task (Chat_ID, Task_Name, Ordering, UserID)
+        values (${chatID.value},${name.value}, ${numberOfTask.value}, ${userID.value})
        """.update
 
-  def delete(chatID: ChatID, numberOfTask: NumberOfTask): Update0 =
+  def delete(
+      chatID: ChatID,
+      numberOfTask: NumberOfTask,
+      userID: UserID
+  ): Update0 =
     sql"""
-         Delete from Task Where Chat_id = ${chatID.value} and Ordering = ${numberOfTask.value}
+         Delete from Task Where Chat_id = ${chatID.value} and Ordering = ${numberOfTask.value} and UserId = ${userID.value}
          """.update
 
-  def update(chatID: ChatID, numberOfTask: NumberOfTask, name: Name): Update0 =
+  def update(
+      chatID: ChatID,
+      numberOfTask: NumberOfTask,
+      name: Name,
+      userID: UserID
+  ): Update0 =
     sql"""
-         Update Task set Task_Name = ${name} Where Chat_id = ${chatID.value} and Ordering = ${numberOfTask.value}
+         Update Task set Task_Name = ${name} Where Chat_id = ${chatID.value} and Ordering = ${numberOfTask.value} and UserId = ${userID.value}
          """.update
 
   def getByChat(chatID: ChatID): Query0[TodoTask] =
     sql"""Select * from Task where CHAT_ID = ${chatID.value}"""
       .query[TodoTask]
 
+  def getTasksByUser(chatID: ChatID, userID: UserID): Query0[TodoTask] =
+    sql""" Select * from Task where CHAT_ID = ${chatID.value} and UserID = ${userID.value}
+         """.query[TodoTask]
+
   def count(chatID: ChatID): Query0[Int] =
     sql"""Select Count(*) from Task where CHAT_ID = ${chatID.value}"""
       .query[Int]
+
 }
